refactor: reorganize filter components into features directory
- Updated import paths for various filter-related components, moving them from the `filters` directory to a new `features/filters` directory for improved structure and clarity.
- This includes changes to imports in layout, sidebar, and filter card components, enhancing code organization and maintainability.
