+++ src/app/(app)/(stacked-layout)/my-filters/page.tsx
@@ -14,7 +14,7 @@ import {
 
 import { getShareToken } from "@/actions/shareTokens";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import { SavedFilters } from "@/components/filters/saved-filters";
+import { SavedFilters } from "@/components/features/filters/components/saved-filters";
 import { MyFilters } from "@/components/my-filters";
 import { MyFiltersHeading } from "@/components/my-filters/my-filters-heading";
 import { ShareHelpDialog } from "@/components/my-filters/shared-filters/share-help-dialog";
