+++ src/components/features/filters/components/filter-items-carousel.tsx
