+++ src/components/features/filters/filter-card/filter-card-header.tsx
@@ -1,9 +1,9 @@
 import type { ConveyorFilterWithAuthor } from "@/types/filter";
 import { CardHeader, CardTitle } from "@/components/ui/card";
-import { BookmarkToggle } from "@/components/filters/filter-card/bookmark-toggle";
-import { FilterCardDescription } from "@/components/filters/filter-card/filter-card-description";
-import { FilterCardMeta } from "@/components/filters/filter-card/filter-card-meta";
-import { ShareButton } from "@/components/filters/filter-card/share-button";
+import { BookmarkToggle } from "@/components/features/filters/filter-card/bookmark-toggle";
+import { FilterCardDescription } from "@/components/features/filters/filter-card/filter-card-description";
+import { FilterCardMeta } from "@/components/features/filters/filter-card/filter-card-meta";
+import { ShareButton } from "@/components/features/filters/filter-card/share-button";
 
 export function FilterCardHeader({
   filter,
